# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SimFull": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "hasDetail": {
            "description": "False once an old practice run is trimmed to its summary",
            "title": "Hasdetail",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "practice",
              "league"
            ],
            "title": "Kind",
            "type": "string"
          },
          "pinned": {
            "title": "Pinned",
            "type": "boolean"
          },
          "season": {
            "title": "Season",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "season",
          "title",
          "summary",
          "pinned",
          "hasDetail",
          "createdAt",
          "updatedAt",
          "version",
          "detail"
        ],
        "title": "SimFull",
        "type": "object"
      },
      "SimIn": {
        "additionalProperties": false,
        "description": "A finished practice draft or a new replay league. `id` is the client's (idempotent).",
        "properties": {
          "detail": {
            "additionalProperties": true,
            "description": "The full run or league (\u2264 256 KB)",
            "title": "Detail",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "practice",
              "league"
            ],
            "title": "Kind",
            "type": "string"
          },
          "season": {
            "title": "Season",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "description": "Small (\u2264 4 KB): what the list shows",
            "title": "Summary",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "season",
          "title",
          "summary",
          "detail"
        ],
        "title": "SimIn",
        "type": "object"
      },
      "SimPin": {
        "properties": {
          "pinned": {
            "title": "Pinned",
            "type": "boolean"
          }
        },
        "required": [
          "pinned"
        ],
        "title": "SimPin",
        "type": "object"
      },
      "SimSummary": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "hasDetail": {
            "description": "False once an old practice run is trimmed to its summary",
            "title": "Hasdetail",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "practice",
              "league"
            ],
            "title": "Kind",
            "type": "string"
          },
          "pinned": {
            "title": "Pinned",
            "type": "boolean"
          },
          "season": {
            "title": "Season",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "season",
          "title",
          "summary",
          "pinned",
          "hasDetail",
          "createdAt",
          "updatedAt",
          "version"
        ],
        "title": "SimSummary",
        "type": "object"
      },
      "SimUpdate": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "summary",
          "detail"
        ],
        "title": "SimUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read-only views over the published daily outputs (D-62), plus users, roles and invites (D-64).",
    "title": "NBA Fantasy Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/me/sims": {
      "get": {
        "operationId": "list_sims_me_sims_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "practice",
                    "league"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SimSummary"
                  },
                  "title": "Response List Sims Me Sims Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sims",
        "tags": [
          "sims"
        ]
      },
      "post": {
        "operationId": "save_sim_me_sims_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Already saved (same id): the stored copy"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimFull"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "too-many-leagues"
          },
          "413": {
            "description": "Over 256 KB"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Sim",
        "tags": [
          "sims"
        ]
      }
    },
    "/me/sims/{sim_id}": {
      "delete": {
        "operationId": "delete_sim_me_sims__sim_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sim_id",
            "required": true,
            "schema": {
              "title": "Sim Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Sim",
        "tags": [
          "sims"
        ]
      },
      "get": {
        "operationId": "get_sim_me_sims__sim_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "sim_id",
            "required": true,
            "schema": {
              "title": "Sim Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimFull"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Sim",
        "tags": [
          "sims"
        ]
      },
      "put": {
        "description": "A replay league as it progresses (weeks kept, roster): compare-and-set on its version.",
        "operationId": "update_league_me_sims__sim_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "sim_id",
            "required": true,
            "schema": {
              "title": "Sim Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimFull"
                }
              }
            },
            "description": "Successful Response"
          },
          "412": {
            "description": "Changed elsewhere: the body's `current` is the stored copy"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "description": "If-Match is required"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update League",
        "tags": [
          "sims"
        ]
      }
    },
    "/me/sims/{sim_id}/pin": {
      "patch": {
        "operationId": "pin_sim_me_sims__sim_id__pin_patch",
        "parameters": [
          {
            "in": "path",
            "name": "sim_id",
            "required": true,
            "schema": {
              "title": "Sim Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimPin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "too-many-pins"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pin Sim",
        "tags": [
          "sims"
        ]
      }
    },
...
  }
}